requirements.txt
ai21==2.3.1
ai21-tokenizer==0.9.1
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
altair==5.3.0
annotated-types==0.6.0
anyio==4.2.0
appdirs==1.4.4
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.3
beautifulsoup4==4.12.3
blinker==1.8.2
Brotli==1.1.0
bs4==0.0.2
build==1.2.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.2.1
crewai==0.11.0
crewai-tools==0.2.6
cryptography==42.0.7
cycler==0.12.1
dataclasses-json==0.6.4
debugpy==1.8.1
decorator==5.1.1
Deprecated==1.2.14
deprecation==2.1.0
distro==1.9.0
docstring-parser==0.15
docx2txt==0.8
embedchain==0.1.102
emoji==2.10.1
exceptiongroup==1.2.0
executing==2.0.1
fairscale==0.4.13
faiss-cpu==1.7.4
fastapi==0.110.3
filelock==3.14.0
filetype==1.2.0
fire==0.5.0
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
fonttools==4.51.0
frozendict==2.4.0
frozenlist==1.4.1
fsspec==2024.5.0
geojson==2.5.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.0
google-auth==2.29.0
google-cloud-aiplatform==1.51.0
google-cloud-bigquery==3.23.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
gptcache==0.1.43
greenlet==3.0.3
groq==0.6.0
grpc-google-iam-v1==0.13.0
grpcio==1.63.0
grpcio-status==1.62.2
h11==0.14.0
HackRequests==1.2
hiq-python==1.1.12
html5lib==1.1
httpcore==1.0.2
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.0
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
importlib_resources==6.4.0
iniconfig==2.0.0
instructor==0.5.2
ipykernel==6.29.4
ipython==8.24.0
isodate==0.6.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
kubernetes==29.0.0
lancedb==0.5.7
langchain==0.1.20
langchain-ai21==0.1.5
langchain-community==0.0.38
langchain-core==0.1.52
langchain-groq==0.1.4
langchain-openai==0.0.2.post1
langchain-text-splitters==0.0.2
langdetect==1.0.9
langsmith==0.1.60
litellm==1.37.16
logzero==1.7.0
lxml==5.1.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.20.2
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
multitasking==0.0.11
mutagen==1.47.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nifpy==2.1.0
nltk==3.8.1
nodeenv==1.8.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnxruntime==1.18.0
openai==1.7.1
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-exporter-otlp-proto-http==1.22.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.45b0
orjson==3.10.3
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pandas==2.2.0
parso==0.8.4
peewee==3.17.1
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
playwright==1.44.0
pluggy==1.5.0
posthog==3.5.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.2
psutil==5.9.8
ptyprocess==0.7.0
pulsar-client==3.5.0
pure-eval==0.2.2
py==1.11.0
py-itree==0.0.19
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycrypto==2.6.1
pycryptodomex==3.20.0
pydantic==2.6.1
pydantic_core==2.16.2
pydeck==0.9.1
pyee==11.1.0
PyGithub==1.59.1
Pygments==2.17.2
PyJWT==2.8.0
pylance==0.9.18
pyllama==0.0.9
PyNaCl==1.5.0
pyotp==2.9.0
pyowm==3.3.0
pyparsing==3.1.2
pypdf==3.17.4
PyPika==0.48.9
pyproject_hooks==1.1.0
pyright==1.1.363
pysbd==0.3.4
PySocks==1.7.1
pytest==8.2.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-iso639==2024.2.7
python-magic==0.4.27
pytils==0.4.1
pytube==15.0.0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
rapidfuzz==3.6.1
ratelimiter==1.2.0.post0
rdflib==7.0.0
rdflib-jsonld==0.6.2
referencing==0.35.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
retry==0.9.2
rich==13.7.0
rpds-py==0.18.1
rsa==4.9
schema==0.7.7
sec-api==1.0.17
selenium==4.21.0
semver==3.0.2
sentencepiece==0.2.0
shapely==2.0.4
simplejson==3.19.2
six==1.16.0
smartapi-python==1.4.8
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.27
stack-data==0.6.3
starlette==0.37.2
streamlit==1.34.0
sympy==1.12
ta==0.11.0
tabulate==0.9.0
tenacity==8.3.0
termcolor==2.4.0
tiktoken==0.5.2
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
tools==0.1.9
toolz==0.12.1
torch==2.3.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
trio==0.25.1
trio-websocket==0.11.1
triton==2.3.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
unstructured==0.10.25
urllib3==2.2.0
uvicorn==0.29.0
uvloop==0.19.0
watchdog==4.0.0
watchfiles==0.21.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
wikipedia==1.4.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
yfinance==0.2.35
youtube-transcript-api==0.6.2
yt-dlp==2023.12.30
zipp==3.17.0
